// src/types/legacy-supabase.ts
export type Json = string | number | boolean | null | { [key: string]: Json | undefined } | Json[];

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5";
  };
  public: {
    Tables: {
// ...
      orders: {
        Row: {
          admin_viewed_at: string | null;
          app_fee: number;
          coupon_code: string | null;
          coupon_id: string | null;
          courier_company: string | null;
          courier_etd: string | null;
          courier_service: string | null;
          created_at: string;
          delivered_at: string | null;
          discount_amount: number;
          id: string;
          notes: string | null;
          order_number: string;
          recipient_name: string;
          recipient_phone: string;
          refund_account_name: string | null;
          refund_account_number: string | null;
          refund_bank_name: string | null;
          shipping_address: string;
          shipping_city: string;
          shipping_cost: number;
          shipping_district: string;
          shipping_insurance: number;
          shipping_lat: number | null;
          shipping_lng: number | null;
          shipping_postal: string;
          shipping_province: string;
          status: Database["public"]["Enums"]["order_status"];
          subtotal: number;
          total: number;
          updated_at: string;
          user_id: string | null;
        };
        Insert: {
          admin_viewed_at?: string | null;
          app_fee?: number;
          coupon_code?: string | null;
          coupon_id?: string | null;
          courier_company?: string | null;
          courier_etd?: string | null;
          courier_service?: string | null;
          created_at?: string;
          delivered_at?: string | null;
          discount_amount?: number;
          id?: string;
          notes?: string | null;
          order_number?: string;
          recipient_name: string;
          recipient_phone: string;
          refund_account_name?: string | null;
          refund_account_number?: string | null;
          refund_bank_name?: string | null;
          shipping_address: string;
          shipping_city: string;
          shipping_cost?: number;
          shipping_district: string;
          shipping_insurance?: number;
          shipping_lat?: number | null;
          shipping_lng?: number | null;
          shipping_postal: string;
          shipping_province: string;
          status?: Database["public"]["Enums"]["order_status"];
          subtotal: number;
          total: number;
          updated_at?: string;
          user_id?: string | null;
        };
        Update: {
          admin_viewed_at?: string | null;
          app_fee?: number;
          coupon_code?: string | null;
          coupon_id?: string | null;
          courier_company?: string | null;
          courier_etd?: string | null;
          courier_service?: string | null;
          created_at?: string;
          delivered_at?: string | null;
          discount_amount?: number;
          id?: string;
          notes?: string | null;
          order_number?: string;
          recipient_name?: string;
          recipient_phone?: string;
          refund_account_name?: string | null;
          refund_account_number?: string | null;
          refund_bank_name?: string | null;
          shipping_address?: string;
          shipping_city?: string;
          shipping_cost?: number;
          shipping_district?: string;
          shipping_insurance?: number;
          shipping_lat?: number | null;
          shipping_lng?: number | null;
          shipping_postal?: string;
          shipping_province?: string;
          status?: Database["public"]["Enums"]["order_status"];
          subtotal?: number;
          total?: number;
          updated_at?: string;
          user_id?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "orders_coupon_id_fkey";
            columns: ["coupon_id"];
            isOneToOne: false;
            referencedRelation: "coupons";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "orders_user_id_fkey";
            columns: ["user_id"];
            isOneToOne: false;
            referencedRelation: "profiles";
            referencedColumns: ["id"];
          },
        ];
      };
      payments: {
        Row: {
          created_at: string;
          expiry_time: string | null;
          gross_amount: number;
          id: string;
          mayar_payment_id: string | null;
          mayar_transaction_id: string | null;
          midtrans_order_id: string;
          midtrans_transaction_id: string | null;
          order_id: string;
          paid_at: string | null;
          payment_code: string | null;
          payment_type: string | null;
          payment_url: string | null;
          pdf_url: string | null;
          provider: string;
          raw_response: Json | null;
          status: Database["public"]["Enums"]["payment_status"];
          updated_at: string;
          va_number: string | null;
        };
        Insert: {
          created_at?: string;
          expiry_time?: string | null;
          gross_amount: number;
          id?: string;
          mayar_payment_id?: string | null;
          mayar_transaction_id?: string | null;
          midtrans_order_id: string;
          midtrans_transaction_id?: string | null;
          order_id: string;
          paid_at?: string | null;
          payment_code?: string | null;
          payment_type?: string | null;
          payment_url?: string | null;
          pdf_url?: string | null;
          provider?: string;
          raw_response?: Json | null;
          status?: Database["public"]["Enums"]["payment_status"];
          updated_at?: string;
          va_number?: string | null;
        };
        Update: {
          created_at?: string;
          expiry_time?: string | null;
          gross_amount?: number;
          id?: string;
          mayar_payment_id?: string | null;
          mayar_transaction_id?: string | null;
          midtrans_order_id?: string;
          midtrans_transaction_id?: string | null;
          order_id?: string;
          paid_at?: string | null;
          payment_code?: string | null;
          payment_type?: string | null;
          payment_url?: string | null;
          pdf_url?: string | null;
          provider?: string;
          raw_response?: Json | null;
          status?: Database["public"]["Enums"]["payment_status"];
          updated_at?: string;
          va_number?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "payments_order_id_fkey";
            columns: ["order_id"];
            isOneToOne: false;
            referencedRelation: "orders";
            referencedColumns: ["id"];
          },
        ];
      };
// ...
    };
// ...
  };
};

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">;

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    keyof DefaultSchema["Enums"] | { schema: keyof DatabaseWithoutInternals },
  EnumName extends (DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never) = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never;